chore: Refactor out common code from BPop ops
