+++ src/server/CMakeLists.txt
@@ -18,7 +18,7 @@ add_library(dragonfly_lib channel_store.cc command_registry.cc
             config_flags.cc conn_context.cc debugcmd.cc dflycmd.cc
             generic_family.cc hset_family.cc json_family.cc search_family.cc
             list_family.cc main_service.cc memory_cmd.cc rdb_load.cc rdb_save.cc replica.cc
-            snapshot.cc script_mgr.cc server_family.cc malloc_stats.cc
+            snapshot.cc script_mgr.cc server_family.cc malloc_stats.cc multi_key_blocking.cc
             set_family.cc stream_family.cc string_family.cc
             zset_family.cc version.cc bitops_family.cc container_utils.cc io_utils.cc
             serializer_commons.cc journal/serializer.cc journal/executor.cc journal/streamer.cc
